requirements.txt
requests
python-dotenv
psycopg2-binary
pandas
SQLAlchemy
streamlit
APScheduler
plotly 